Refine index.d.ts and Update package.json Version
This commit addresses and resolves issues previously encountered with the 
index.d.ts file. The modifications include commenting out the export type 
of 'containerFactory' and 'coreFactoryStarter', as well as the default 
export of 'containerFactory'. In addition, a new 'anyType' has been declared 
and exported as the default.

In the package.json, version 1.8.45 has been upgraded to 1.8.47 and the 
'dsl-framework' dependency version has been updated from 1.8.23 to 1.8.26. 
Further, changes have been made to the "main" and "module" fields, now 
pointing to "./dist/src/index.js". The previous "exports" object, which 
included types and default exports, has been replaced with these updates.

These corrections have enhanced the functionality of the 'demeter-di' 
package, making it a significant point in the project's timeline. The 
changes will ensure that the index.d.ts file operates efficiently and align 
with the updated version of the package.